split connection file
