+++ h3/src/connection/connection.rs
@@ -1,12 +1,7 @@
-use std::{
-    convert::TryFrom,
-    marker::PhantomData,
-    sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard},
-};
+use std::marker::PhantomData;
 
-use bytes::{Buf, Bytes, BytesMut};
-use futures_util::{future, Future};
-use http::HeaderMap;
+use bytes::Buf;
+use futures_util::future;
 use tracing::warn;
 
 use crate::{
@@ -14,67 +9,17 @@ use crate::{
     frame::FrameStream,
     proto::{
         frame::{Frame, PayloadLen, SettingId, Settings},
-        headers::Header,
-        stream::{StreamId, StreamType},
+        stream::StreamType,
         varint::VarInt,
     },
-    qpack,
     quic::{self, CloseCon, SendStream as _},
     stream::{self, AcceptRecvStream, AcceptedRecvStream},
 };
 
-#[doc(hidden)]
-pub struct SharedState {
-    // maximum size for a header we send
-    pub peer_max_field_section_size: u64,
-    // connection-wide error, concerns all RequestStreams and drivers
-    pub error: Option<Error>,
-    // Has the connection received a GoAway frame? If so, this StreamId is the last
-    // we're willing to accept. This lets us finish the requests or pushes that were
-    // already in flight when the graceful shutdown was initiated.
-    pub closing: Option<StreamId>,
-    /// Stream ID of the last accepted Stream
-    pub last_accepted_stream: Option<StreamId>,
-}
-
-#[derive(Clone)]
-#[doc(hidden)]
-pub struct SharedStateRef(Arc<RwLock<SharedState>>);
-
-impl SharedStateRef {
-    pub fn read(&self, panic_msg: &'static str) -> RwLockReadGuard<SharedState> {
-        self.0.read().expect(panic_msg)
-    }
-
-    pub fn write(&self, panic_msg: &'static str) -> RwLockWriteGuard<SharedState> {
-        self.0.write().expect(panic_msg)
-    }
-}
-
-impl Default for SharedStateRef {
-    fn default() -> Self {
-        Self(Arc::new(RwLock::new(SharedState {
-            peer_max_field_section_size: VarInt::MAX.0,
-            error: None,
-            closing: None,
-            last_accepted_stream: None,
-        })))
-    }
-}
-
-pub trait ConnectionState {
-    fn shared_state(&self) -> &SharedStateRef;
-
-    /// checks if there has been an connection error somewhere and returns the Connection  Error.
-    /// If there was no Connection Error it returns the given Error
-    fn maybe_conn_err<E: Into<Error>>(&self, err: E) -> Error {
-        if let Some(ref e) = self.shared_state().0.read().unwrap().error {
-            e.clone()
-        } else {
-            err.into()
-        }
-    }
-}
+use super::{
+    connection_state::{ConnectionState, SharedStateRef},
+    control_stream_send_handler::ControlStreamSendHandler,
+};
 
 pub trait HasQuicConnection<B: Buf> {
     type Conn: quic::CloseCon;
@@ -194,49 +139,13 @@ pub(crate) async fn start_connection<C: quic::Connection<B> + quic::CloseCon, B:
     Ok((conn_inner, control_send_handler))
 }
 
-pub struct ControlStreamSendHandler<S, B>
-where
-    S: crate::quic::SendStream<B>,
-    B: Buf,
-{
-    pub(super) shared: SharedStateRef,
-    control_send: S,
-    buf: PhantomData<B>,
-}
-
-impl<S, B> ControlStreamSendHandler<S, B>
-where
-    S: crate::quic::SendStream<B>,
-    B: Buf,
-{
-    /// Initiate graceful shutdown, accepting `max_streams` potentially in-flight streams
-    pub async fn shutdown(&mut self, max_streams: usize) -> Result<(), Error> {
-        let max_id = self
-            .shared
-            .read("failed to read last_accepted_stream")
-            .last_accepted_stream
-            .map(|id| id + max_streams)
-            .unwrap_or_else(StreamId::first_request);
-
-        self.shared.write("graceful shutdown").closing = Some(max_id);
-
-        //= https://www.rfc-editor.org/rfc/rfc9114#section-3.3
-        //# When either endpoint chooses to close the HTTP/3
-        //# connection, the terminating endpoint SHOULD first send a GOAWAY frame
-        //# (Section 5.2) so that both endpoints can reliably determine whether
-        //# previously sent frames have been processed and gracefully complete or
-        //# terminate any necessary remaining tasks.
-        stream::write(&mut self.control_send, Frame::Goaway(max_id)).await
-    }
-}
-
 /// Accepts Unidirectional streams like the
 pub struct UnidirectionalStreamAcceptHandler<C, B>
 where
     C: quic::Connection<B> + quic::CloseCon,
     B: Buf,
 {
-    pub(super) shared: SharedStateRef,
+    pub(crate) shared: SharedStateRef,
     conn: C,
     control_recv: Option<FrameStream<C::RecvStream, B>>,
     decoder_recv: Option<AcceptedRecvStream<C::RecvStream, B>>,
@@ -246,27 +155,16 @@ where
     got_peer_settings: bool,
 }
 
-pub trait H3FutureExecutor
-{
-    fn run_future(&self, future: impl Future + Send + 'static);
-}
-
 impl<C, B> UnidirectionalStreamAcceptHandler<C, B>
 where
     C: quic::Connection<B> + quic::CloseCon,
     B: Buf,
 {
-    pub async fn accept_recv<Fut: H3FutureExecutor>(
-        &mut self,
-        future_executor: Fut,
-    ) -> Result<(), Error> {
+    pub async fn accept_recv(&mut self) -> Result<(), Error> {
+        // TODO: put this somewhere else.
         if let Some(ref e) = self.shared.read("poll_accept_request").error {
             return Err(e.clone());
         }
-
-        future_executor.run_future(async {
-            println!("HI---------------");
-        });
 
         let mut accept_stream = AcceptRecvStream::new(self.conn.accept_recv().await?);
         accept_stream.receive_type().await?;
@@ -308,7 +206,8 @@ where
         Ok(())
     }
 
-    pub async fn control(&mut self) -> Result<Frame<PayloadLen>, Error> {
+    pub async fn handle_connection_state(&mut self) -> Result<Frame<PayloadLen>, Error> {
+        // TODO: put this somewhere else.
         if let Some(ref e) = self.shared.read("poll_accept_request").error {
             return Err(e.clone());
         }
@@ -526,31 +425,6 @@ where
     code.with_reason(reason.as_ref(), crate::error::ErrorLevel::ConnectionError)
 }
 
-pub struct RequestStream<S, B> {
-    pub(super) stream: FrameStream<S, B>,
-    pub(super) trailers: Option<Bytes>,
-    pub(super) conn_state: SharedStateRef,
-    pub(super) max_field_section_size: u64,
-    send_grease_frame: bool,
-}
-
-impl<S, B> RequestStream<S, B> {
-    pub fn new(
-        stream: FrameStream<S, B>,
-        max_field_section_size: u64,
-        conn_state: SharedStateRef,
-        grease: bool,
-    ) -> Self {
-        Self {
-            stream,
-            conn_state,
-            max_field_section_size,
-            trailers: None,
-            send_grease_frame: grease,
-        }
-    }
-}
-
 impl<C, B> ConnectionState for UnidirectionalStreamAcceptHandler<C, B>
 where
     B: Buf,
@@ -570,234 +444,5 @@ where
 
     fn get_conn(&mut self) -> &mut Self::Conn {
         &mut self.conn
-    }
-}
-
-impl<S, B> ConnectionState for RequestStream<S, B> {
-    fn shared_state(&self) -> &SharedStateRef {
-        &self.conn_state
-    }
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::RecvStream,
-{
-    /// Receive some of the request body.
-    pub async fn recv_data(&mut self) -> Result<Option<impl Buf>, Error> {
-        if !self.stream.has_data() {
-            let frame = future::poll_fn(|cx| self.stream.poll_next(cx))
-                .await
-                .map_err(|e| self.maybe_conn_err(e))?;
-            match frame {
-                Some(Frame::Data { .. }) => (),
-                Some(Frame::Headers(encoded)) => {
-                    self.trailers = Some(encoded);
-                    return Ok(None);
-                }
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-4.1
-                //# Receipt of an invalid sequence of frames MUST be treated as a
-                //# connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.3
-                //# Receiving a
-                //# CANCEL_PUSH frame on a stream other than the control stream MUST be
-                //# treated as a connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.4
-                //# If an endpoint receives a SETTINGS frame on a different
-                //# stream, the endpoint MUST respond with a connection error of type
-                //# H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.6
-                //# A client MUST treat a GOAWAY frame on a stream other than
-                //# the control stream as a connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.7
-                //# The MAX_PUSH_ID frame is always sent on the control stream.  Receipt
-                //# of a MAX_PUSH_ID frame on any other stream MUST be treated as a
-                //# connection error of type H3_FRAME_UNEXPECTED.
-                Some(_) => return Err(Code::H3_FRAME_UNEXPECTED.into()),
-                None => return Ok(None),
-            }
-        }
-
-        let data = future::poll_fn(|cx| self.stream.poll_data(cx))
-            .await
-            .map_err(|e| self.maybe_conn_err(e))?;
-        Ok(data)
-    }
-
-    /// Receive trailers
-    pub async fn recv_trailers(&mut self) -> Result<Option<HeaderMap>, Error> {
-        let mut trailers = if let Some(encoded) = self.trailers.take() {
-            encoded
-        } else {
-            let frame = future::poll_fn(|cx| self.stream.poll_next(cx))
-                .await
-                .map_err(|e| self.maybe_conn_err(e))?;
-            match frame {
-                Some(Frame::Headers(encoded)) => encoded,
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-4.1
-                //# Receipt of an invalid sequence of frames MUST be treated as a
-                //# connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.3
-                //# Receiving a
-                //# CANCEL_PUSH frame on a stream other than the control stream MUST be
-                //# treated as a connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.4
-                //# If an endpoint receives a SETTINGS frame on a different
-                //# stream, the endpoint MUST respond with a connection error of type
-                //# H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.6
-                //# A client MUST treat a GOAWAY frame on a stream other than
-                //# the control stream as a connection error of type H3_FRAME_UNEXPECTED.
-
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-7.2.7
-                //# The MAX_PUSH_ID frame is always sent on the control stream.  Receipt
-                //# of a MAX_PUSH_ID frame on any other stream MUST be treated as a
-                //# connection error of type H3_FRAME_UNEXPECTED.
-                Some(_) => return Err(Code::H3_FRAME_UNEXPECTED.into()),
-                None => return Ok(None),
-            }
-        };
-
-        if !self.stream.is_eos() {
-            // Get the trailing frame
-            let trailing_frame = future::poll_fn(|cx| self.stream.poll_next(cx))
-                .await
-                .map_err(|e| self.maybe_conn_err(e))?;
-
-            if trailing_frame.is_some() {
-                // if it's not unknown or reserved, fail.
-                return Err(Code::H3_FRAME_UNEXPECTED.into());
-            }
-        }
-
-        let qpack::Decoded { fields, .. } =
-            match qpack::decode_stateless(&mut trailers, self.max_field_section_size) {
-                //= https://www.rfc-editor.org/rfc/rfc9114#section-4.2.2
-                //# An HTTP/3 implementation MAY impose a limit on the maximum size of
-                //# the message header it will accept on an individual HTTP message.
-
-                // Todo: Send response with status 431, see spec
-                Err(qpack::DecoderError::HeaderTooLong(cancel_size)) => {
-                    return Err(Error::header_too_big(
-                        cancel_size,
-                        self.max_field_section_size,
-                    ))
-                }
-                Ok(decoded) => decoded,
-                Err(e) => return Err(e.into()),
-            };
-
-        Ok(Some(Header::try_from(fields)?.into_fields()))
-    }
-
-    pub fn stop_sending(&mut self, err_code: Code) {
-        self.stream.stop_sending(err_code);
-    }
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::SendStream<B>,
-    B: Buf,
-{
-    /// Send some data on the response body.
-    pub async fn send_data(&mut self, buf: B) -> Result<(), Error> {
-        let frame = Frame::Data(buf);
-
-        stream::write(&mut self.stream, frame)
-            .await
-            .map_err(|e| self.maybe_conn_err(e))?;
-        Ok(())
-    }
-
-    /// Send a set of trailers to end the request.
-    pub async fn send_trailers(&mut self, trailers: HeaderMap) -> Result<(), Error> {
-        //= https://www.rfc-editor.org/rfc/rfc9114#section-4.2
-        //= type=TODO
-        //# Characters in field names MUST be
-        //# converted to lowercase prior to their encoding.
-        let mut block = BytesMut::new();
-
-        let mem_size = qpack::encode_stateless(&mut block, Header::trailer(trailers))?;
-        let max_mem_size = self
-            .conn_state
-            .read("send_trailers shared state read")
-            .peer_max_field_section_size;
-
-        //= https://www.rfc-editor.org/rfc/rfc9114#section-4.2.2
-        //# An implementation that
-        //# has received this parameter SHOULD NOT send an HTTP message header
-        //# that exceeds the indicated size, as the peer will likely refuse to
-        //# process it.
-        if mem_size > max_mem_size {
-            return Err(Error::header_too_big(mem_size, max_mem_size));
-        }
-        stream::write(&mut self.stream, Frame::Headers(block.freeze()))
-            .await
-            .map_err(|e| self.maybe_conn_err(e))?;
-
-        Ok(())
-    }
-
-    /// Stops an stream with an error code
-    pub fn stop_stream(&mut self, code: Code) {
-        self.stream.reset(code.into());
-    }
-
-    pub async fn finish(&mut self) -> Result<(), Error> {
-        if self.send_grease_frame {
-            // send a grease frame once per Connection
-            stream::write(&mut self.stream, Frame::Grease)
-                .await
-                .map_err(|e| self.maybe_conn_err(e))?;
-            self.send_grease_frame = false;
-        }
-        future::poll_fn(|cx| self.stream.poll_ready(cx))
-            .await
-            .map_err(|e| self.maybe_conn_err(e))?;
-        future::poll_fn(|cx| self.stream.poll_finish(cx))
-            .await
-            .map_err(|e| self.maybe_conn_err(e))
-    }
-}
-
-impl<S, B> RequestStream<S, B>
-where
-    S: quic::BidiStream<B>,
-    B: Buf,
-{
-    pub(crate) fn split(
-        self,
-    ) -> (
-        RequestStream<S::SendStream, B>,
-        RequestStream<S::RecvStream, B>,
-    ) {
-        let (send, recv) = self.stream.split();
-
-        (
-            RequestStream {
-                stream: send,
-                trailers: None,
-                conn_state: self.conn_state.clone(),
-                max_field_section_size: 0,
-                send_grease_frame: self.send_grease_frame,
-            },
-            RequestStream {
-                stream: recv,
-                trailers: self.trailers,
-                conn_state: self.conn_state,
-                max_field_section_size: self.max_field_section_size,
-                send_grease_frame: self.send_grease_frame,
-            },
-        )
     }
 }
